requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
cryptography==41.0.3
python-dotenv==1.0.0
PyJWT==2.8.0